feat: rename About app to About Alex
