+++ components/apps/alex/SafetyNote.tsx
