+++ pages/apps/alex.jsx
@@ -1,0 +1,8 @@
+import dynamic from 'next/dynamic';
+
+const PageAlex = dynamic(() => import('../../apps/alex'), {
+  ssr: false,
+  loading: () => <p>Loading...</p>,
+});
+
+export default PageAlex;
